Fix incorrect comparison of sentry-cli version
Prior to this commit the code would incorrectly report that version 1.0.0
of the sentry-cli was less than the required version of 0.25.0. The issue
was that “1.0.0” was converted to 100 and “0.25.0” was converted to 250,
and thus comparing `sentry_cli_version` to `required_version` did not
give the expected result.
Leveraging `Gem::Version`s support for parsing and comparing versions
fixes the problem.