fix: Prevent buffers deadlock by merging stdout and stderr
> You should be careful to avoid deadlocks. Since pipes are fixed length buffers, Open3.popen3(“prog”) {|i, o, e, t| o.read } deadlocks if the program generates too much output on stderr. You should read stdout and stderr simultaneously (using threads or IO.select). However, if you don't need stderr output, you can use Open3.popen2. If merged stdout and stderr output is not a problem, you can use Open3.popen2e. If you really need stdout and stderr output as separate strings, you can consider Open3.capture3.
— https://www.rubydoc.info/stdlib/open3/Open3.popen3

This changes how output is printed, but for the better IMO. As this way, we preserve the output ordering of sentry-cli, instead of bundling stdout/stderr together. Starting `sentry-cli` v2, all logs are going go `stderr`, so `--verbose` will control whether it's available inside `stdout_and_stderr` or not.

Also we want to solely rely on exit codes, and not use `/error/ =~` as it's just unnecessary.

Fixes https://github.com/getsentry/sentry-fastlane-plugin/issues/129